New GPIO Library

+++ src/comp/cylinder.rs
@@ -38,7 +38,7 @@ impl crate::math::MathActor for Cylinder
 
 impl crate::meas::SimpleMeas for Cylinder 
 {
-    fn init_meas(&mut self, pin_meas : u16) {
+    fn init_meas(&mut self, pin_meas : u8) {
         self.ctrl.init_meas(pin_meas)
     }
 }
