+++ src/tests/servo.rs
@@ -3,8 +3,8 @@ use std::io;
 use crate::Time;
 use crate::ctrl::pwm::PWMOutput;
 
-const PIN_1 : u16 = 13;
-const PIN_2 : u16 = 23;
+const PIN_1 : u8 = 13;
+const PIN_2 : u8 = 23;
 
 #[test]
 fn pwm_servo() {
